refactor(core): split checkout cart use case
- move checkout finalization out of the cart mutation module
- keep the public use-case export stable through the index barrel
- reduce cart.ts from 315 to 226 LOC and cyclomatic total from 43 to 31
